Small updates and fixing tests + refactor
Small changes, refactored somethings as well as starting writing the predict input.
