+++ Carbon/__tests__/Progress/GetData.test.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react-native';
-import GetData from '../../../../navigation/screens/Home/GetData';
+import GetData from '../../navigation/screens/Progress/GetData.js';
 /*
     This needs to be updated with more tests
     It's currently having issues with fetching
