+++ Carbon/__tests__/calculations/calcBottles.test.js
@@ -3,5 +3,5 @@ import calcBottles from "../../calculations/calcBottles.js"
 test("calcBottles", () =>
 {
     ret = calcBottles(30);
-    expect(ret).equalsTo(Math.round(.343 * 30));
+    expect(ret).toBe(Math.round(.343 * 30));
 });
