+++ Carbon/__tests__/calculations/food_calculations/calcCheese.test.js
@@ -3,5 +3,5 @@ import calcCheese from "../../../calculations/food_calculations/calcCheese.js"
 test("calcCheese", async() =>
 {
     ret = await calcCheese(3)
-    expect(ret).equalsTo(Math.round(3 * 16.6792));
+    expect(ret).toBe(Math.round(3 * 16.6792));
 })
