+++ Carbon/navigation/screens/Progress/GetData.js
@@ -5,15 +5,14 @@ import { API_URL } from "../../../config/Api";
 */
 const GetData = async() => {
   try {
-      const user_id = 323;
+      const user_id = 323; //needs to be not hard coded
       //const url = "http://{YOURLOCALIPHERE}:3000/api/userEmissions?user_id="+ user_id; //for local hosting and testing 
 
-      const url = API_URL + `userEmissions?user_id=${user_id}`; //for the database
+      const url = API_URL + `userEmissions?user_id=${user_id}`; //for the database n
       console.log("fetching data from " + url); //log to see if if it works
     
       const response = await fetch(url); //wait for response
       const data = await response.json(); //get the data we need
-      console.log(data);
       return data
     }
     catch (error) {
